Fix deblocking masks and indexing on larger images

+++ src/loopfilter.h
@@ -39,7 +39,7 @@
 // TODO: * Compute q_thr/side_thr from seg_ids in filter.
 //       * Add a flag to shift down q_thr/side_thr for sub_pu_edge
 #define decl_loopfilter_sb_fn(name) \
-void (name)(pixel *dst, ptrdiff_t stride, const uint32_t *mask, \
+void (name)(pixel *dst, ptrdiff_t stride, const uint64_t *mask, \
             unsigned q_thr, unsigned side_thr, \
             const Av1FilterLUT *lut, int w HIGHBD_DECL_SUFFIX)
 typedef decl_loopfilter_sb_fn(*loopfilter_sb_fn);
